some further attempts but it wont work for now (comments)
So if you use a flex server then the way to add users is to log into the postgres database and select into a special function.  I will have to consider if we want to do that.  If it is not a flex server, what we're doing here should work